# public/api-docs/mobile-app/mobile-app-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SCD Wellness \u2014 Mobile App API",
        "description": "REST API for the SCD Wellness Flutter app. Two roles: **patient** (mobile) and **doctor** (web dashboard). Protected routes require `Authorization: Bearer {token}`.",
        "contact": {
            "email": "api@scdwellness.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://statrawebapp-f7gaa7bghfczhmf7.centralus-01.azurewebsites.net",
            "description": "Production"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local dev"
        }
    ],
    "paths": {
...
        "/api/v1/doctor/patients": {
            "get": {
                "tags": [
                    "Doctor Patients"
                ],
                "summary": "List all assigned patients",
                "operationId": "cd5f02476ad10d2271613fba61395c47",
                "responses": {
                    "200": {
                        "description": "Patient list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/doctor/patients/{id}": {
            "get": {
                "tags": [
                    "Doctor Patients"
                ],
                "summary": "Get full patient detail",
                "operationId": "4246622dabd740a7a3ce40fa206f8a3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Doctor Patients"
                ],
                "summary": "Update patient assignment and clinical info \u2014 403 for staff role",
                "operationId": "87e3fc7d6b0292e5220bb87382a28944",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "assigned_doctor_id": {
                                        "description": "Reassign attending doctor",
                                        "type": "integer",
                                        "example": 3,
                                        "nullable": true
                                    },
                                    "assigned_nurse_id": {
                                        "description": "Reassign attending nurse",
                                        "type": "integer",
                                        "example": 5,
                                        "nullable": true
                                    },
                                    "hospital_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "ward": {
                                        "type": "string",
                                        "example": "Haematology Ward A",
                                        "nullable": true
                                    },
                                    "admitted_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-01",
                                        "nullable": true
                                    },
                                    "genotype": {
                                        "type": "string",
                                        "example": "SS",
                                        "nullable": true
                                    },
                                    "blood_type": {
                                        "type": "string",
                                        "example": "O+",
                                        "nullable": true
                                    },
                                    "gender": {
                                        "type": "string",
                                        "nullable": true,
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ]
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1998-03-14",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Patient updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Staff role cannot update patients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful."
                    },
                    "data": {
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PaginatedResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 20
                            },
                            "total": {
                                "type": "integer",
                                "example": 84
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 5
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Invalid credentials."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}